+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ What's new in version 0.9
   introduced in version 0.7. (:issue:`248`)
 - Ability to disable totals plot with `totals_plot_elements=0`. (:issue:`246`)
 - Ability to set totals y axis label (:issue:`243`)
+- Added ``max_subset_rank`` to get only n most populous subsets.
 
 What's new in version 0.8
 -------------------------